фикс рантайма

+++ code/game/alternate_appearance.dm
@@ -11,7 +11,7 @@ GLOBAL_LIST_EMPTY(active_alternate_appearances)
 
 		alternate_appearances.Remove(possible_key)
 		qdel(alternate_appearance)
-		if(!alternate_appearances.len)
+		if(alternate_appearances && !alternate_appearances.len)
 			qdel(alternate_appearances)
 
 		break
